+++ BoxOffice/Controller/SceneDelegate.swift
@@ -18,7 +18,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
         guard let windowScene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: windowScene)
-        let viewController = ViewController(dataSource: BoxOffice())
+        let viewController = ViewController(dataSource: BoxOfficeDTO())
         window?.rootViewController = viewController
         window?.makeKeyAndVisible()
     }
